MB-18955: Compress temporary index files
The temporary files that are created during indexing weren't compressed.
This lead to huge disk usage. When those files are compressed, it even
leads to an overall better indexing performance.

The compressed files consist of chunks. It's always a prefix with
the size of the compressed data, followed  by the data itself.

Change-Id: I9ca7a800e9c79e11e23b216ef24d332eea43019e